refactor: Change the office parsing code to dynamic import to reduce the number of files loaded on the landing page refactor: Change the system instructions and attachment area components to dynamic loading
