+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "gemini-next-chat",
-  "version": "1.8.1",
+  "version": "1.8.2",
   "private": true,
   "author": "Amery2010 <amery@xiangfa.org>",
   "license": "MIT",
