+++ src-tauri/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "gemini-next-chat"
-version = "1.8.1"
+version = "1.8.2"
 description = "Your private Gemini application."
 authors = ["Amery2010<amery@xiangfa.org>"]
 license = "MIT"
