+++ src-tauri/tauri.conf.json
@@ -36,7 +36,7 @@
   },
   "productName": "gemini-next-chat",
   "mainBinaryName": "gemini-next-chat",
-  "version": "1.8.1",
+  "version": "1.8.2",
   "identifier": "com.u14.app",
   "plugins": {},
   "app": {
